refactor(GUI): extract MainPage from `lib/gui/app.js`
`lib/gui/app.js` contains a lot of code that should have been split long
ago. This PR extracts the "main page" logic into an actual page
component in `lib/gui/pages`.
