+++ lib/gui/app.js
@@ -26,51 +26,35 @@ var angular = require('angular');
 
 /* eslint-enable no-var */
 
-const _ = require('lodash');
 const Store = require('./models/store');
 
 const app = angular.module('Etcher', [
   require('angular-ui-router'),
   require('angular-ui-bootstrap'),
-  require('angular-moment'),
-  require('angular-middle-ellipses'),
   require('angular-if-state'),
-  require('angular-seconds-to-date'),
 
   // Etcher modules
-  require('./modules/drive-scanner'),
-  require('./modules/image-writer'),
   require('./modules/analytics'),
 
   // Models
   require('./models/selection-state'),
-  require('./models/settings'),
-  require('./models/supported-formats'),
-  require('./models/drives'),
   require('./models/flash-state'),
 
   // Components
-  require('./components/progress-button/progress-button'),
-  require('./components/drive-selector/drive-selector'),
   require('./components/svg-icon/svg-icon'),
   require('./components/update-notifier/update-notifier'),
-  require('./components/tooltip-modal/tooltip-modal'),
 
   // Pages
+  require('./pages/main/main'),
   require('./pages/finish/finish'),
   require('./pages/settings/settings'),
 
   // OS
-  require('./os/notification/notification'),
   require('./os/window-progress/window-progress'),
   require('./os/open-external/open-external'),
-  require('./os/dropzone/dropzone'),
-  require('./os/dialog/dialog'),
 
   // Utils
-  require('./utils/path/path'),
-  require('./utils/manifest-bind/manifest-bind'),
-  require('./utils/byte-size/byte-size')
+  require('./utils/manifest-bind/manifest-bind')
 ]);
 
 app.run((AnalyticsService, UpdateNotifierService, SelectionStateModel) => {
@@ -121,219 +105,6 @@ app.run((AnalyticsService, OSWindowProgressService, FlashStateModel) => {
   });
 });
 
-app.config(($stateProvider, $urlRouterProvider) => {
+app.config(($urlRouterProvider) => {
   $urlRouterProvider.otherwise('/main');
-
-  $stateProvider
-    .state('main', {
-      url: '/main',
-      controller: 'AppController as app',
-      templateUrl: './partials/main.html'
-    });
-});
-
-app.controller('AppController', function(
-  $state,
-  DriveScannerService,
-  SelectionStateModel,
-  FlashStateModel,
-  SettingsModel,
-  SupportedFormatsModel,
-  DrivesModel,
-  ImageWriterService,
-  AnalyticsService,
-  DriveSelectorService,
-  UpdateNotifierService,
-  TooltipModalService,
-  OSWindowProgressService,
-  OSNotificationService,
-  OSDialogService,
-  OSOpenExternalService
-) {
-  this.formats = SupportedFormatsModel;
-  this.selection = SelectionStateModel;
-  this.drives = DrivesModel;
-  this.writer = ImageWriterService;
-  this.state = FlashStateModel;
-  this.settings = SettingsModel;
-  this.tooltipModal = TooltipModalService;
-
-  this.handleError = (error) => {
-
-    // This particular error is handled by the alert ribbon
-    // on the main application page.
-    if (error.code === 'ENOSPC') {
-      AnalyticsService.logEvent('Drive ran out of space');
-      return;
-    }
-
-    OSDialogService.showError(error);
-
-    // Also throw it so it gets displayed in DevTools
-    // and its reported by TrackJS.
-    throw error;
-  };
-
-  // This catches the case where the user enters
-  // the settings screen when a flash finished
-  // and goes back to the main screen with the back button.
-  if (!FlashStateModel.isFlashing()) {
-
-    this.selection.clear({
-
-      // Preserve image, in case there is one, otherwise
-      // we revert the behaviour of "Use same image".
-      preserveImage: true
-
-    });
-  }
-
-  DriveScannerService.start();
-
-  DriveScannerService.on('error', this.handleError);
-
-  DriveScannerService.on('drives', (drives) => {
-    this.drives.setDrives(drives);
-
-    if (_.isEmpty(drives)) {
-      DriveSelectorService.close();
-    }
-  });
-
-  this.selectImage = (image) => {
-    if (!SupportedFormatsModel.isSupportedImage(image.path)) {
-      OSDialogService.showError('Invalid image', `${image.path} is not a supported image type.`);
-      AnalyticsService.logEvent('Invalid image', image);
-      return;
-    }
-
-    this.selection.setImage(image);
-    AnalyticsService.logEvent('Select image', _.omit(image, 'logo'));
-  };
-
-  this.openImageUrl = () => {
-    const imageUrl = this.selection.getImageUrl();
-
-    if (imageUrl) {
-      OSOpenExternalService.open(imageUrl);
-    }
-  };
-
-  this.openImageSelector = () => {
-    return OSDialogService.selectImage().then((image) => {
-
-      // Avoid analytics and selection state changes
-      // if no file was resolved from the dialog.
-      if (!image) {
-        return;
-      }
-
-      this.selectImage(image);
-    }).catch(this.handleError);
-  };
-
-  this.selectDrive = (drive) => {
-    if (!drive) {
-      return;
-    }
-
-    this.selection.setDrive(drive.device);
-
-    AnalyticsService.logEvent('Select drive', {
-      device: drive.device
-    });
-  };
-
-  this.openDriveSelector = () => {
-    DriveSelectorService.open()
-      .then(this.selectDrive)
-      .catch(this.handleError);
-  };
-
-  this.reselectImage = () => {
-    if (FlashStateModel.isFlashing()) {
-      return;
-    }
-
-    // Reselecting an image automatically
-    // de-selects the current drive, if any.
-    // This is made so the user effectively
-    // "returns" to the first step.
-    this.selection.clear();
-
-    this.openImageSelector();
-    AnalyticsService.logEvent('Reselect image');
-  };
-
-  this.reselectDrive = () => {
-    if (FlashStateModel.isFlashing()) {
-      return;
-    }
-
-    this.openDriveSelector();
-    AnalyticsService.logEvent('Reselect drive');
-  };
-
-  this.restartAfterFailure = () => {
-    this.selection.clear({
-      preserveImage: true
-    });
-
-    FlashStateModel.resetState();
-    AnalyticsService.logEvent('Restart after failure');
-  };
-
-  this.wasLastFlashSuccessful = () => {
-    const flashResults = FlashStateModel.getFlashResults();
-
-    if (_.get(flashResults, 'cancelled', false)) {
-      return true;
-    }
-
-    return _.get(flashResults, 'passedValidation', true);
-  };
-
-  this.flash = (image, drive) => {
-
-    if (FlashStateModel.isFlashing()) {
-      return;
-    }
-
-    // Stop scanning drives when flashing
-    // otherwise Windows throws EPERM
-    DriveScannerService.stop();
-
-    AnalyticsService.logEvent('Flash', {
-      image: image,
-      device: drive.device
-    });
-
-    return this.writer.flash(image, drive).then(() => {
-      const results = FlashStateModel.getFlashResults();
-
-      if (results.cancelled) {
-        return;
-      }
-
-      if (results.passedValidation) {
-        OSNotificationService.send('Success!', 'Your flash is complete');
-        AnalyticsService.logEvent('Done');
-        $state.go('success');
-      } else {
-        OSNotificationService.send('Oops!', 'Looks like your flash has failed');
-        AnalyticsService.logEvent('Validation error');
-      }
-    })
-    .catch((error) => {
-
-      if (error.type === 'check') {
-        AnalyticsService.logEvent('Validation error');
-      } else {
-        AnalyticsService.logEvent('Flash error');
-      }
-
-      this.handleError(error);
-    })
-    .finally(OSWindowProgressService.clear);
-  };
 });
