+++ lib/gui/pages/main/templates/main.tpl.html
@@ -1,0 +1,126 @@
+<div class="row around-xs">
+  <div class="col-xs">
+    <div class="box text-center" os-dropzone="main.selectImage($file)">
+      <svg-icon class="center-block" path="{{ main.selection.getImageLogo() || '../../../assets/image.svg' }}"></svg-icon>
+      <span class="icon-caption">SELECT IMAGE</span>
+      <span class="badge space-top-medium">1</span>
+
+      <div class="space-vertical-large">
+        <div ng-hide="main.selection.hasImage()">
+          <button class="btn btn-primary btn-brick" ng-click="main.openImageSelector()">Select image</button>
+
+          <p class="step-footer">
+            {{ ::main.formats.getAllExtensions().slice(0, 3).join(', ') }}, and
+            <span class="step-footer-underline"
+              uib-tooltip="{{ main.formats.getAllExtensions().slice(3).join(', ') }}">many more</span>
+          </p>
+        </div>
+        <div ng-if="main.selection.hasImage()">
+          <div ng-click="main.openImageUrl()"
+            ng-bind="main.selection.getImageName() || main.selection.getImagePath() | basename | middleEllipses:25"></div>
+
+          <button class="btn btn-link step-tooltip"
+            ng-click="main.tooltipModal.show({
+              title: 'IMAGE FILE NAME',
+              message: main.selection.getImagePath()
+            })">SHOW IN FULL</button>
+
+          <button class="btn btn-link step-footer"
+            ng-click="main.reselectImage()"
+            ng-hide="main.state.isFlashing()">Change</button>
+        </div>
+      </div>
+    </div>
+  </div>
+
+  <div class="col-xs">
+    <div class="box text-center relative">
+      <div class="step-border-left" ng-disabled="!main.selection.hasImage()"></div>
+      <div class="step-border-right" ng-disabled="!main.selection.hasImage() || !main.selection.hasDrive()"></div>
+
+      <svg-icon class="center-block"
+        path="../../../assets/drive.svg"
+        ng-disabled="!main.selection.hasImage()"></svg-icon>
+      <span class="icon-caption"
+        ng-disabled="!main.selection.hasImage()">SELECT DRIVE</span>
+
+      <span class="badge space-top-medium" ng-disabled="!main.selection.hasImage()">2</span>
+
+      <div class="space-vertical-large">
+        <div ng-hide="main.selection.hasDrive()">
+
+          <div ng-show="main.drives.hasAvailableDrives() || !main.selection.hasImage()">
+            <button class="btn btn-primary btn-brick"
+              ng-disabled="!main.selection.hasImage()"
+              ng-click="main.openDriveSelector()">Select drive</button>
+          </div>
+
+          <div ng-hide="main.drives.hasAvailableDrives() || !main.selection.hasImage()">
+            <button class="btn btn-danger btn-brick">Connect a drive</button>
+          </div>
+
+        </div>
+        <div ng-show="main.selection.hasDrive()">
+          <div ng-class="{
+            soft: !main.selection.hasImage()
+          }">{{ main.selection.getDrive().name }} - {{ main.selection.getDrive().size | gigabyte | number:1 }} GB</div>
+          <button class="btn btn-link step-footer"
+            ng-click="main.reselectDrive()"
+            ng-hide="main.state.isFlashing()">Change</button>
+        </div>
+      </div>
+    </div>
+  </div>
+
+  <div class="col-xs">
+    <div class="box text-center">
+      <svg-icon class="center-block"
+        path="../../../assets/flash.svg"
+        ng-disabled="!main.selection.hasImage() || !main.selection.hasDrive()"></svg-icon>
+      <span class="icon-caption"
+        ng-disabled="!main.selection.hasImage() || !main.selection.hasDrive()">FLASH IMAGE</span>
+
+      <span class="badge space-top-medium" ng-disabled="!main.selection.hasImage() || !main.selection.hasDrive()">3</span>
+
+      <div class="space-vertical-large">
+        <progress-button class="btn-brick"
+          percentage="main.state.getFlashState().percentage"
+          striped="{{ main.state.getFlashState().type == 'check' }}"
+          ng-attr-active="{{ main.state.isFlashing() }}"
+          ng-show="main.wasLastFlashSuccessful()"
+          ng-click="main.flash(main.selection.getImagePath(), main.selection.getDrive())"
+          ng-disabled="!main.selection.hasImage() || !main.selection.hasDrive()">
+            <span ng-show="main.state.getFlashState().percentage == 100 && main.state.isFlashing()">Finishing...</span>
+            <span ng-show="main.state.getFlashState().percentage == 0 && !main.state.isFlashing()">Flash!</span>
+            <span ng-show="main.state.getFlashState().percentage == 0 && main.state.isFlashing() && !main.state.getFlashState().speed">Starting...</span>
+            <span ng-show="main.state.getFlashState().speed && main.state.getFlashState().percentage != 100 && main.state.getFlashState().type != 'check'"
+              ng-bind="main.state.getFlashState().percentage + '% '"></span>
+            <span ng-show="main.state.getFlashState().speed && main.state.getFlashState().percentage != 100 && main.state.getFlashState().type == 'check'"
+              ng-bind="main.state.getFlashState().percentage + '% Validating...'"></span>
+        </progress-button>
+
+        <div class="alert-ribbon alert-warning" ng-class="{ 'alert-ribbon--open': !main.wasLastFlashSuccessful() }">
+          <span class="glyphicon glyphicon-warning-sign"></span>
+          <span ng-show="main.state.getFlashResults().errorCode === 'ENOSPC'">
+            Not enough space on the drive.<br>Please insert larger one and <button class="btn btn-link" ng-click="main.restartAfterFailure()">try again</button>
+          </span>
+          <span ng-show="main.state.getFlashResults().errorCode !== 'ENOSPC' && main.settings.get('validateWriteOnSuccess')">
+            Your removable drive may be corrupted.<br>Try inserting a different one and <button class="btn btn-link" ng-click="main.restartAfterFailure()">press "retry"</button>
+          </span>
+          <span ng-show="main.state.getFlashResults().errorCode !== 'ENOSPC' && !main.settings.get('validateWriteOnSuccess')">
+            Oops, seems something went wrong. Click <button class="btn btn-link" ng-click="main.restartAfterFailure()">here</button> to retry
+          </span>
+        </div>
+
+        <button class="btn btn-warning btn-brick" ng-hide="main.wasLastFlashSuccessful()" ng-click="main.restartAfterFailure()">
+          <span class="glyphicon glyphicon-repeat"></span> Retry
+        </button>
+
+        <p class="step-footer step-footer-split" ng-show="main.state.getFlashState().speed && main.state.getFlashState().percentage != 100">
+          <span>ETA: {{ main.state.getFlashState().eta | secondsToDate | amDateFormat:'m[m]ss[s]' }}</span>
+          <span ng-bind="main.state.getFlashState().speed.toFixed(2) + ' MB/s'"></span>
+        </p>
+      </div>
+    </div>
+  </div>
+</div>
