--- lib/gui/partials/main.html
@@ -1,126 +1,0 @@
-<div class="row around-xs">
-  <div class="col-xs">
-    <div class="box text-center" os-dropzone="app.selectImage($file)">
-      <svg-icon class="center-block" path="{{ app.selection.getImageLogo() || '../../../assets/image.svg' }}"></svg-icon>
-      <span class="icon-caption">SELECT IMAGE</span>
-      <span class="badge space-top-medium">1</span>
-
-      <div class="space-vertical-large">
-        <div ng-hide="app.selection.hasImage()">
-          <button class="btn btn-primary btn-brick" ng-click="app.openImageSelector()">Select image</button>
-
-          <p class="step-footer">
-            {{ ::app.formats.getAllExtensions().slice(0, 3).join(', ') }}, and
-            <span class="step-footer-underline"
-              uib-tooltip="{{ app.formats.getAllExtensions().slice(3).join(', ') }}">many more</span>
-          </p>
-        </div>
-        <div ng-if="app.selection.hasImage()">
-          <div ng-click="app.openImageUrl()"
-            ng-bind="app.selection.getImageName() || app.selection.getImagePath() | basename | middleEllipses:25"></div>
-
-          <button class="btn btn-link step-tooltip"
-            ng-click="app.tooltipModal.show({
-              title: 'IMAGE FILE NAME',
-              message: app.selection.getImagePath()
-            })">SHOW IN FULL</button>
-
-          <button class="btn btn-link step-footer"
-            ng-click="app.reselectImage()"
-            ng-hide="app.state.isFlashing()">Change</button>
-        </div>
-      </div>
-    </div>
-  </div>
-
-  <div class="col-xs">
-    <div class="box text-center relative">
-      <div class="step-border-left" ng-disabled="!app.selection.hasImage()"></div>
-      <div class="step-border-right" ng-disabled="!app.selection.hasImage() || !app.selection.hasDrive()"></div>
-
-      <svg-icon class="center-block"
-        path="../../../assets/drive.svg"
-        ng-disabled="!app.selection.hasImage()"></svg-icon>
-      <span class="icon-caption"
-        ng-disabled="!app.selection.hasImage()">SELECT DRIVE</span>
-
-      <span class="badge space-top-medium" ng-disabled="!app.selection.hasImage()">2</span>
-
-      <div class="space-vertical-large">
-        <div ng-hide="app.selection.hasDrive()">
-
-          <div ng-show="app.drives.hasAvailableDrives() || !app.selection.hasImage()">
-            <button class="btn btn-primary btn-brick"
-              ng-disabled="!app.selection.hasImage()"
-              ng-click="app.openDriveSelector()">Select drive</button>
-          </div>
-
-          <div ng-hide="app.drives.hasAvailableDrives() || !app.selection.hasImage()">
-            <button class="btn btn-danger btn-brick">Connect a drive</button>
-          </div>
-
-        </div>
-        <div ng-show="app.selection.hasDrive()">
-          <div ng-class="{
-            soft: !app.selection.hasImage()
-          }">{{ app.selection.getDrive().name }} - {{ app.selection.getDrive().size | gigabyte | number:1 }} GB</div>
-          <button class="btn btn-link step-footer"
-            ng-click="app.reselectDrive()"
-            ng-hide="app.state.isFlashing()">Change</button>
-        </div>
-      </div>
-    </div>
-  </div>
-
-  <div class="col-xs">
-    <div class="box text-center">
-      <svg-icon class="center-block"
-        path="../../../assets/flash.svg"
-        ng-disabled="!app.selection.hasImage() || !app.selection.hasDrive()"></svg-icon>
-      <span class="icon-caption"
-        ng-disabled="!app.selection.hasImage() || !app.selection.hasDrive()">FLASH IMAGE</span>
-
-      <span class="badge space-top-medium" ng-disabled="!app.selection.hasImage() || !app.selection.hasDrive()">3</span>
-
-      <div class="space-vertical-large">
-        <progress-button class="btn-brick"
-          percentage="app.state.getFlashState().percentage"
-          striped="{{ app.state.getFlashState().type == 'check' }}"
-          ng-attr-active="{{ app.state.isFlashing() }}"
-          ng-show="app.wasLastFlashSuccessful()"
-          ng-click="app.flash(app.selection.getImagePath(), app.selection.getDrive())"
-          ng-disabled="!app.selection.hasImage() || !app.selection.hasDrive()">
-            <span ng-show="app.state.getFlashState().percentage == 100 && app.state.isFlashing()">Finishing...</span>
-            <span ng-show="app.state.getFlashState().percentage == 0 && !app.state.isFlashing()">Flash!</span>
-            <span ng-show="app.state.getFlashState().percentage == 0 && app.state.isFlashing() && !app.state.getFlashState().speed">Starting...</span>
-            <span ng-show="app.state.getFlashState().speed && app.state.getFlashState().percentage != 100 && app.state.getFlashState().type != 'check'"
-              ng-bind="app.state.getFlashState().percentage + '% '"></span>
-            <span ng-show="app.state.getFlashState().speed && app.state.getFlashState().percentage != 100 && app.state.getFlashState().type == 'check'"
-              ng-bind="app.state.getFlashState().percentage + '% Validating...'"></span>
-        </progress-button>
-
-        <div class="alert-ribbon alert-warning" ng-class="{ 'alert-ribbon--open': !app.wasLastFlashSuccessful() }">
-          <span class="glyphicon glyphicon-warning-sign"></span>
-          <span ng-show="app.state.getFlashResults().errorCode === 'ENOSPC'">
-            Not enough space on the drive.<br>Please insert larger one and <button class="btn btn-link" ng-click="app.restartAfterFailure()">try again</button>
-          </span>
-          <span ng-show="app.state.getFlashResults().errorCode !== 'ENOSPC' && app.settings.get('validateWriteOnSuccess')">
-            Your removable drive may be corrupted.<br>Try inserting a different one and <button class="btn btn-link" ng-click="app.restartAfterFailure()">press "retry"</button>
-          </span>
-          <span ng-show="app.state.getFlashResults().errorCode !== 'ENOSPC' && !app.settings.get('validateWriteOnSuccess')">
-            Oops, seems something went wrong. Click <button class="btn btn-link" ng-click="app.restartAfterFailure()">here</button> to retry
-          </span>
-        </div>
-
-        <button class="btn btn-warning btn-brick" ng-hide="app.wasLastFlashSuccessful()" ng-click="app.restartAfterFailure()">
-          <span class="glyphicon glyphicon-repeat"></span> Retry
-        </button>
-
-        <p class="step-footer step-footer-split" ng-show="app.state.getFlashState().speed && app.state.getFlashState().percentage != 100">
-          <span>ETA: {{ app.state.getFlashState().eta | secondsToDate | amDateFormat:'m[m]ss[s]' }}</span>
-          <span ng-bind="app.state.getFlashState().speed.toFixed(2) + ' MB/s'"></span>
-        </p>
-      </div>
-    </div>
-  </div>
-</div>
